requirements.txt
gradio
tiktoken
python-docx
PyPDF2
openai
ipywidgets
transformers
